moved addressDocketSwitchRuling to docketSwitchSlice.js
