+++ client/app/queue/docketSwitch/docketSwitchSlice.js
@@ -71,8 +71,23 @@ export const createDocketSwitchRulingTask = createAsyncThunk(
   }
 );
 
-export const completeDocketSwitchMailTask = createAsyncThunk(
-  'tasks/completeDocketSwitchRecommendationTask',
+export const addressDocketSwitchRuling = createAsyncThunk(
+  'tasks/addressDocketSwitchRuling',
+  async (data) => {
+    try {
+      const res = await ApiUtil.post('/docket_switches/address_ruling', { data });
+
+      return res.body;
+    } catch (error) {
+      console.error('Error creating task', error);
+      throw error;
+    }
+  }
+);
+
+// This should likely move to a more generic task slice when it exists
+export const completeTask = createAsyncThunk(
+  'tasks/complete',
   async ({ taskId }, { dispatch }) => {
     try {
       const res = await ApiUtil.patch(`/tasks/${taskId}`, {
